Move `InputChanges` to work package
