+++ subprojects/execution/src/main/java/org/gradle/internal/execution/history/changes/InputChangesInternal.java
@@ -16,8 +16,8 @@
 
 package org.gradle.internal.execution.history.changes;
 
-import org.gradle.api.execution.incremental.InputChanges;
 import org.gradle.internal.change.Change;
+import org.gradle.work.InputChanges;
 
 public interface InputChangesInternal extends InputChanges {
     Iterable<Change> getAllFileChanges();
